emotion/requirements.txt
Flask==2.0.1
opencv-python-headless==4.6.0.66
mediapipe==0.10.15
Werkzeug==2.0.3
matplotlib==3.4.3
numpy==1.21.0
gunicorn==20.1.0
google-oauth2-client==1.0.0
google-api-python-client==2.100.0
flask-cors==4.0.0